Delete unused files and update dependencies
