+++ .env
@@ -1,1 +1,13 @@
 DATABASE_URL="mongodb+srv://oehab558:omarOMAR12345@cluster0.8twkswc.mongodb.net/school"
+CLOUDINARY_URL=cloudinary://962275345124399:r36MMXHKeJoEns1EuwxfKOmqByU@dvtnyioug
+NEXT_PUBLIC_CLOUDINARY_CLOUD_NAME=dvtnyioug
+
+CLOUDINARY_API_KEY=962275345124399
+CLOUDINARY_API_SECRET=r36MMXHKeJoEns1EuwxfKOmqByU
+CLOUDINARY_NAME=dvtnyioug
+
+UPLOADTHING_SECRET=sk_live_2d1e4257093de3dba5159db71c1129ecfde0fae104c907e2054247be18fdf9a8
+UPLOADTHING_APP_ID=lyn9lc91ou
+
+UPLOADTHING_SECRET=sk_live_2d1e4257093de3dba5159db71c1129ecfde0fae104c907e2054247be18fdf9a8
+UPLOADTHING_APP_ID=lyn9lc91ou
