+++ app/api/uploadthing/route.ts
@@ -1,0 +1,9 @@
+import { createNextRouteHandler } from "uploadthing/next";
+
+import { ourFileRouter } from "./core";
+
+// Export routes for Next App Router
+export const { GET, POST } = createNextRouteHandler({
+  
+  router: ourFileRouter,
+});
