+++ app/esteathan/layout.jsx
@@ -1,0 +1,33 @@
+import { Inter } from "next/font/google";
+import "./globals.css";
+import Link from "next/link";
+import "../../node_modules/@fortawesome/fontawesome-free/css/all.min.css";
+import "../../node_modules/bootstrap/dist/css/bootstrap.rtl.min.css";
+import "@uploadthing/react/styles.css";
+// Uncomment the line below if you need Bootstrap's JavaScript functionality
+// import '../node_modules/bootstrap/dist/js/bootstrap.bundle.min.js';
+// import { NextSSRPlugin } from "@uploadthing/react/next-ssr-plugin";
+// import { extractRouterConfig } from "uploadthing/server";
+ 
+// import { ourFileRouter } from "./api/uploadthing/core";
+
+
+const inter = Inter({ subsets: ["latin"] });
+
+export const metadata = {
+  title: "إستئذان",
+  description: "Generated by create next app",
+};
+
+export default function RootLayout({ children }) {
+  return (
+    <html lang="ar" dir="rtl">
+      <body
+        className={` d-flex justify-content-center reqbody flex-wrap ${inter.className}`}
+      >
+        
+        {children}
+      </body>
+    </html>
+  );
+}
