--- app/layout.jsx
@@ -1,22 +1,0 @@
-import { Inter } from "next/font/google";
-import "./globals.css";
-import Link from "next/link";
-import "../node_modules/@fortawesome/fontawesome-free/css/all.min.css";
-import "../node_modules/bootstrap/dist/css/bootstrap.rtl.min.css";
-// Uncomment the line below if you need Bootstrap's JavaScript functionality
-// import '../node_modules/bootstrap/dist/js/bootstrap.bundle.min.js';
-
-const inter = Inter({ subsets: ["latin"] });
-
-export const metadata = {
-  title: "إستئذان",
-  description: "Generated by create next app",
-};
-
-export default function RootLayout({ children }) {
-  return (
-    <html lang="ar" dir="rtl">
-      <body className={` d-flex justify-content-center reqbody flex-wrap ${inter.className}`}>{children}</body>
-    </html>
-  );
-}
