+++ package.json
@@ -12,12 +12,17 @@
   "dependencies": {
     "@fortawesome/fontawesome-free": "^6.4.2",
     "@prisma/client": "^5.6.0",
+    "@uploadthing/react": "^6.0.2",
     "bootstrap": "^5.3.2",
+    "cloudinary": "^1.41.0",
+    "formidable": "^3.5.1",
     "next": "13.5.5",
+    "next-cloudinary": "^5.13.0",
     "node-fetch": "^3.3.2",
     "prisma": "^5.6.0",
     "react": "^18",
-    "react-dom": "^18"
+    "react-dom": "^18",
+    "uploadthing": "^6.1.0"
   },
   "devDependencies": {
     "@types/node": "^20.10.2",
