+++ tsconfig.json
@@ -27,7 +27,7 @@
     ".next/types/**/*.ts",
     "**/*.ts",
     "**/*.tsx"
-  ],
+, "lib/upload.js", "lib/uploadthing.js", "app/api/uploadthing/core.js", "app/api/uploadthing/route.js", "app/api/uploading/route.ts"  ],
   "exclude": [
     "node_modules"
   ]
